jercept/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "jercept"
version = "1.2.0"
description = "The authorization layer for AI agents. IBAC: Intent-Based Access Control."
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
keywords = ["ai", "security", "agents", "prompt-injection", "authorization", "ibac", "jercept"]
classifiers = [
  "Development Status :: 5 - Production/Stable",
  "Intended Audience :: Developers",
  "Topic :: Security",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
]
dependencies = [
  "openai>=1.30.0",
  "httpx>=0.27.0",
  "pyyaml>=6.0",
  "click>=8.1.0",
]

[project.scripts]
jercept = "jercept.cli:main"

[project.optional-dependencies]
openai-agents = ["openai-agents>=0.0.3"]
langchain     = ["langchain>=0.2.0", "langchain-openai>=0.1.0"]
crewai        = ["crewai>=0.1.0"]
mcp           = ["mcp>=1.0.0"]
autogen       = ["pyautogen>=0.2.0"]
llamaindex    = ["llama-index>=0.10.0"]
anthropic     = ["anthropic>=0.25.0"]
gemini        = ["google-generativeai>=0.5.0"]
dev           = ["pytest>=8.0", "pytest-asyncio>=0.23", "pytest-mock>=3.14"]
all           = [
  "openai-agents>=0.0.3",
  "langchain>=0.2.0", "langchain-openai>=0.1.0",
  "crewai>=0.1.0",
  "mcp>=1.0.0",
  "pyautogen>=0.2.0",
  "llama-index>=0.10.0",
  "anthropic>=0.25.0",
  "google-generativeai>=0.5.0",
]

[project.urls]
Homepage      = "https://jercept.com"
Repository    = "https://github.com/jercept/jercept"
Documentation = "https://docs.jercept.com"

[tool.setuptools.packages.find]
where   = ["."]
include = ["jercept*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths    = ["tests"]
